requirements.txt
# Web framework
Flask==3.0.2

# Password hashing
bcrypt==4.1.2
argon2-cffi==23.1.0

# Utilities
requests==2.31.0
pyotp==2.9.0

# Logging and analysis
pandas==2.2.2